# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book Search",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/acquisition/downloads": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Start",
        "operationId": "start_api_acquisition_downloads_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Listing",
        "operationId": "listing_api_acquisition_downloads_get",
        "parameters": [
          {
            "name": "selection_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selection Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/acquisition/downloads/{attempt_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Detail",
        "operationId": "detail_api_acquisition_downloads__attempt_id__get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "downloads"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_acquisition_downloads__attempt_id__delete",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/acquisition/downloads/{attempt_id}/recheck": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Recheck",
        "operationId": "recheck_api_acquisition_downloads__attempt_id__recheck_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/acquisition/downloads/{attempt_id}/repair-preview": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Repair Preview",
        "operationId": "repair_preview_api_acquisition_downloads__attempt_id__repair_preview_get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/acquisition/downloads/{attempt_id}/repairs": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Repair Download",
        "operationId": "repair_download_api_acquisition_downloads__attempt_id__repairs_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttemptView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "AttemptPage"
      },
      "AttemptView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "selection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selection Id"
          },
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "work_title": {
            "type": "string",
            "title": "Work Title"
          },
          "release_title": {
            "type": "string",
            "title": "Release Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "external_may_exist": {
            "type": "boolean",
            "title": "External May Exist"
          },
          "can_recheck": {
            "type": "boolean",
            "title": "Can Recheck"
          },
          "can_cancel": {
            "type": "boolean",
            "title": "Can Cancel"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "inspection_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Id"
          },
          "fulfillment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FulfillmentView"
              },
              {
                "type": "null"
              }
            ]
          },
          "repair": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairView"
              },
              {
                "type": "null"
              }
            ]
          },
          "can_repair": {
            "type": "boolean",
            "title": "Can Repair"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "selection_id",
          "operation_id",
          "state",
          "work_title",
          "release_title",
          "message",
          "external_may_exist",
          "can_recheck",
          "can_cancel",
          "progress",
          "inspection_id",
          "fulfillment",
          "repair",
          "can_repair"
        ],
        "title": "AttemptView"
      },
...
      "FulfillmentView": {
        "properties": {
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          },
          "basis": {
            "type": "string",
            "title": "Basis"
          },
          "available_now": {
            "type": "boolean",
            "title": "Available Now"
          }
        },
        "type": "object",
        "required": [
          "confirmed_at",
          "basis",
          "available_now"
        ],
        "title": "FulfillmentView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepairInput": {
        "properties": {
          "revision": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "RepairInput"
      },
      "RepairPreview": {
        "properties": {
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "changes"
        ],
        "title": "RepairPreview"
      },
      "RepairView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "message",
          "changes",
          "created_at",
          "applied_at"
        ],
        "title": "RepairView"
      },
...
      "StartInput": {
        "properties": {
          "selection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selection Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "selection_id"
        ],
        "title": "StartInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}